address feedback and use withFlowtoggle instead of withEditToggle
