+++ packages/bodiless-components/src/filterHOC.tsx
@@ -1,8 +1,11 @@
 import {
-  withoutProps, ifToggledOn, ifToggledOff, TagType,
+  withoutProps,
+  ifToggledOn,
+  ifToggledOff,
+  TagType,
 } from '@bodiless/core';
 import { replaceWith } from '@bodiless/fclasses';
-import { difference, flowRight } from 'lodash';
+import { difference, flowRight, differenceWith } from 'lodash';
 import { toJS } from 'mobx';
 import useTagsAccessors from './TagButton/TagModel';
 
@@ -24,7 +27,14 @@ const useToggleByTags = <P extends object>({
   console.log('item tags', toJS(tags));
   console.log('differnece', difference(selectedTags, tags));
   // const tags = itemTags.map(t => t.name);
-  return difference(selectedTags, tags).length === 0;
+  return (
+    differenceWith(
+      selectedTags,
+      tags,
+      (selectedTag, itemTag) => selectedTag.name === itemTag.name,
+    ).length === 0
+  );
+  // return difference(selectedTags, tags).length === 0;
 };
 
 const ifTagsSelected = ifToggledOn(useToggleByTags);
