+++ packages/bodiless-core/src/index.ts
@@ -34,10 +34,9 @@ import {
   withLocalContextMenu,
   withContextActivator,
   withoutProps,
-  ifToggledOff,
-  ifToggledOn,
 } from './hoc';
-import { ifEditable, ifReadOnly, withEditToggle } from './withEditToggle';
+import { ifToggledOff, ifToggledOn, withFlowToggle } from './withFlowToggle';
+import { ifEditable, ifReadOnly, useEditToggle } from './withEditToggle';
 import { TMenuOption, PageEditContextInterface } from './PageEditContext/types';
 import { EditButtonOptions } from './Types/EditButtonTypes';
 import { TMenuOptionGetter } from './Types/PageContextProviderTypes';
@@ -82,7 +81,6 @@ export {
   DefaultContentNode,
   ifEditable,
   ifReadOnly,
-  withEditToggle,
   withoutProps,
   ActivateOnEffectProvider,
   withActivateOnEffect,
@@ -93,6 +91,8 @@ export {
   withResetButton,
   ifToggledOff,
   ifToggledOn,
+  withFlowToggle,
+  useEditToggle,
 };
 
 export type Bodiless<P, Q> = (C: ComponentType<P> | string) => ComponentType<Q>;
